+++ packages/api/src/promise/simple-test.spec.ts
@@ -1,0 +1,8 @@
+// Copyright 2017-2025 @polkadot/api authors & contributors
+// SPDX-License-Identifier: Apache-2.0
+
+describe('simple test', (): void => {
+  it('should pass', (): void => {
+    expect(true).toBe(true);
+  });
+});