+++ packages/api/test/simple-test.spec.ts
@@ -1,0 +1,10 @@
+// Copyright 2017-2025 @polkadot/api authors & contributors
+// SPDX-License-Identifier: Apache-2.0
+
+/// <reference types="@polkadot/dev-test/globals.d.ts" />
+
+describe('simple test', (): void => {
+  it('should pass', (): void => {
+    expect(true).toBe(true);
+  });
+});